Refactor main.py into multiple sub-files
Divide the `main.py` file into multiple sub-files for better organization and maintainability.

* **UI Components**: Add `ui.py` to contain all PyQt6 UI components and related functions. Move the `TextEditor` class and its methods from `main.py` to `ui.py`. Import necessary modules and functions in `ui.py`. Update the `TextEditor` class to work independently in `ui.py`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/SchBenedikt/Text-Editor?shareId=XXXX-XXXX-XXXX-XXXX).
